web_version/backend/requirements.txt
# Created: 2026-04-21 00:00:00
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.26.0
